fixing grammar issues and refactor

+++ src/Command/ModelCommand.php
@@ -157,13 +157,29 @@ class ".$createdFile." extends ResourceCollection{
 		}
 	}
 
+	private function alreadyHave(string $createdFile,string $createdOption){
+		echo $createdFile . " ".$createdOption." is already created".PHP_EOL;
+	}
+
+	private function success(string $createdFile,string $createdOption){
+		echo $createdFile . " ".$createdOption." is created successfully".PHP_EOL;
+	}
+
+	private function wrongCommand(){
+		echo "You type wrong command".PHP_EOL;
+	}
+
 	public function run(string $dir,array $argv){
-		try{
-			if(count($argv)==3 && ($argv[1]==$this->modelCommandLine || $argv[1]==$this->observerCommandLine || $argv[1]==$this->resourceCommandLine  ) ){
-				$command=$argv[1];
-				$createdOption=$this->checkOption($command);
-				$defaulFolder=$this->checkPath($command);
-				$baseDir=$dir.'/'.$defaulFolder;
+
+		if(count($argv)==3 && ($argv[1]==$this->modelCommandLine || $argv[1]==$this->observerCommandLine || $argv[1]==$this->resourceCommandLine  ) ){
+			$command=$argv[1];
+			$createdOption=$this->checkOption($command);
+			$defaulFolder=$this->checkPath($command);
+			$baseDir=$dir.'/'.$defaulFolder;
+			if(substr($argv[2], -1)=='/'){
+				return $this->wrongCommand();
+			}
+			try {
 				if(!is_dir($baseDir)){
 					$createdFolder=NULL;
 					$basefolder=explode('/', $defaulFolder);
@@ -176,17 +192,24 @@ class ".$createdFile." extends ResourceCollection{
 				}
 				$inputFile=explode('/',$argv[2]);
 				$count=count($inputFile);
-				if($count==1){
+
+				if($count==1 && $inputFile[0]!==NULL && !file_exists($baseDir.'/'.$inputFile[0].'.php') ){
 					$createdFile=$inputFile[0];
-					if(!file_exists($baseDir.'/'.$createdFile.'.php')){
-						fopen($baseDir.'/'.$createdFile.'.php', 'w') or die('Unable to create model');
+					fopen($baseDir.'/'.$createdFile.'.php', 'w') or die('Unable to create '.$createdOption);
 						$createdFileContent=$this->checkContent($command,$defaulFolder,$createdFile);
 						file_put_contents($baseDir.'/'.$createdFile.'.php', $createdFileContent);
-						echo $createdFile . " ".$createdOption." is created successfully".PHP_EOL;
-					}elseif(file_exists($baseDir . '/'.$createdFile.'.php')){
-						echo $createdFile . " ".$createdOption." is already created".PHP_EOL;
-					}
-				}else{
+						return $this->success($createdFile,$createdOption);
+				
+				}elseif($count==1 && $inputFile[0]!==NULL && file_exists($baseDir . '/'.$inputFile[0].'.php') ){
+					$createdFile=$inputFile[0];
+				
+					return $this->alreadyHave($createdFile,$createdOption);
+				
+				}elseif($count>1 && file_exists($baseDir.'/'. implode('/', $inputFile) . '.php' ) ){
+					$createdFile=implode('/',$inputFile);
+					return $this->alreadyHave($createdFile,$createdOption);
+				
+				}elseif($count>1 && !file_exists($baseDir .'/'. implode('/', $inputFile) . '.php' ) ){
 					$createdFile=$inputFile[$count-1];
 					unset($inputFile[$count-1]);
 					$currentFolder=NULL;
@@ -198,18 +221,18 @@ class ".$createdFile." extends ResourceCollection{
 							mkdir($currentFolder);
 						}
 					}
-					if($currentFolder!==NULL && !file_exists($currentFolder.'/'.$createdFile.'.php') ){
-						fopen($currentFolder.'/'.$createdFile.'.php', 'w') or die('Unable to create model');
+
+					fopen($currentFolder.'/'.$createdFile.'.php', 'w') or die('Unable to create '.$createdOption);
 						$createdFileContent=$this->checkContent($command,$newCreatedFolder,$createdFile);
 						file_put_contents($currentFolder.'/'.$createdFile.'.php', $createdFileContent);
-						echo $createdFile ." ".$createdOption." is created successfully".PHP_EOL;
-					}elseif(file_exists($currentFolder.'/'.$createdFile.'.php')){
-						echo $createdFile . " ".$createdOption." is already created".PHP_EOL;
-					}
+						return $this->success($createdFile,$createdOption);
 				}
+			} catch (Exception $e) {
+				echo "You can't create ". $createdFile . " " . $createdOption.PHP_EOL;
 			}
-		}catch(Exception $e){
-			echo "You can't created Model".PHP_EOL;
+
+		}else{
+			return $this->wrongCommand();
 		}
 	}
